+++ lib/passwordless.rb
@@ -12,8 +12,17 @@ require "passwordless/token_digest"
 module Passwordless
   extend Configurable
 
+  LOCK = Mutex.new
+
   def self.context
-    @context ||= Context.new
+    return @context if @context
+
+    # Routes are lazy loaded in Rails 8 so we need to load them to populate Context#resources.
+    Rails.application.try(:reload_routes_unless_loaded)
+
+    LOCK.synchronize do
+      @context ||= Context.new
+    end
   end
 
   def self.add_resource(resource, controller:, **defaults)