+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2023_06_21_085550) do
+ActiveRecord::Schema[8.0].define(version: 2023_06_21_085550) do
   create_table "admins", force: :cascade do |t|
     t.string "email"
     t.datetime "created_at", null: false
@@ -38,5 +38,4 @@ ActiveRecord::Schema[7.1].define(version: 2023_06_21_085550) do
     t.datetime "updated_at", null: false
     t.index ["email"], name: "index_users_on_email", unique: true
   end
-
 end